Add user registration form and update login features
This commit introduces a new user registration form (formRegistro.jsp) and enhances login functionalities. A doGet method is added in AltaUsuario class to redirect to the registration form. Furthermore, the login features in inicioSesion.jsp are updated to accommodate the new changes. The MensajeBean class is also updated with getters and a new method for custom messages.