Update user registration link and create new User Service
This commit makes several changes in dealing with user registration and login processes. Firstly, it updates the href link in the user registration call-to-action in the inicioSesion.jsp file. Secondly, a new UsuarioService.java file is created to handle the registration process. This includes parsing incoming parameters and organizing them into a new Usuario (User) object. Lastly, the AltaUsuario.java servlet is updated to use the newly created UsuarioService, for both handling GET requests (redirecting to registration form) and POST requests (processing the registration).